fix(telecom.sms): rename option general option to callback option
ref: MANAGER-10745
