+++ packages/manager/modules/sms/src/sms/options/callback/telecom-sms-options-callback.html
@@ -1,11 +1,11 @@
-<section class="telecom-sms-options-manage">
+<section class="telecom-sms-options-callback">
     <header>
         <tuc-section-back-link
             data-tuc-section-back-link-to-state="sms.service.options"
             data-tuc-section-back-link-title="{{:: 'sms_section_back_link' | translate }}"
         >
         </tuc-section-back-link>
-        <h1 data-translate="sms_options_manage_title"></h1>
+        <h1 data-translate="sms_options_callback_title"></h1>
     </header>
 
     <main role="main">
@@ -16,20 +16,20 @@
                 <header class="widget-presentation-header">
                     <h2
                         class="widget-presentation-title"
-                        data-translate="sms_options_manage_information_title"
+                        data-translate="sms_options_callback_information_title"
                     ></h2>
                 </header>
 
-                <p data-translate="sms_options_manage_information_text"></p>
+                <p data-translate="sms_options_callback_information_text"></p>
                 <p
-                    data-translate="sms_options_manage_information_types_available"
+                    data-translate="sms_options_callback_information_types_available"
                 ></p>
                 <ul>
                     <li
-                        data-translate="sms_options_manage_information_type_callback"
+                        data-translate="sms_options_callback_information_type_callback"
                     ></li>
                     <li
-                        data-translate="sms_options_manage_information_type_stop_callback"
+                        data-translate="sms_options_callback_information_type_stop_callback"
                     ></li>
                 </ul>
             </div>
@@ -38,46 +38,46 @@
                 <header class="widget-presentation-header">
                     <h2
                         class="widget-presentation-title"
-                        data-translate="sms_options_manage_configure_title"
+                        data-translate="sms_options_callback_configure_title"
                     ></h2>
                 </header>
 
                 <div
                     class="text-center"
-                    data-ng-if="TelecomSmsOptionsManageCtrl.loading.init"
+                    data-ng-if="TelecomSmsOptionsCallbackCtrl.loading.init"
                 >
                     <oui-spinner></oui-spinner>
                 </div>
 
-                <dl data-ng-if="!TelecomSmsOptionsManageCtrl.loading.init">
+                <dl data-ng-if="!TelecomSmsOptionsCallbackCtrl.loading.init">
                     <dt
-                        data-translate="sms_options_manage_configure_callback_label"
+                        data-translate="sms_options_callback_configure_callback_label"
                     ></dt>
                     <dd>
                         <span
-                            data-ng-if="TelecomSmsOptionsManageCtrl.service.callBack"
-                            data-ng-bind="TelecomSmsOptionsManageCtrl.service.callBack"
+                            data-ng-if="TelecomSmsOptionsCallbackCtrl.service.callBack"
+                            data-ng-bind="TelecomSmsOptionsCallbackCtrl.service.callBack"
                         >
                         </span>
                         <span
-                            data-ng-if="!TelecomSmsOptionsManageCtrl.service.callBack"
-                            data-translate="sms_options_manage_configure_callback_empty"
+                            data-ng-if="!TelecomSmsOptionsCallbackCtrl.service.callBack"
+                            data-translate="sms_options_callback_configure_callback_empty"
                         >
                         </span>
                     </dd>
                     <dt
                         class="mt-3"
-                        data-translate="sms_options_manage_configure_stop_callback_label"
+                        data-translate="sms_options_callback_configure_stop_callback_label"
                     ></dt>
                     <dd>
                         <span
-                            data-ng-if="TelecomSmsOptionsManageCtrl.service.stopCallBack"
-                            data-ng-bind="TelecomSmsOptionsManageCtrl.service.stopCallBack"
+                            data-ng-if="TelecomSmsOptionsCallbackCtrl.service.stopCallBack"
+                            data-ng-bind="TelecomSmsOptionsCallbackCtrl.service.stopCallBack"
                         >
                         </span>
                         <span
-                            data-ng-if="!TelecomSmsOptionsManageCtrl.service.stopCallBack"
-                            data-translate="sms_options_manage_configure_callback_empty"
+                            data-ng-if="!TelecomSmsOptionsCallbackCtrl.service.stopCallBack"
+                            data-translate="sms_options_callback_configure_callback_empty"
                         >
                         </span>
                     </dd>
@@ -86,8 +86,8 @@
                 <button
                     type="button"
                     class="btn btn-primary"
-                    data-ng-if="!TelecomSmsOptionsManageCtrl.loading.init"
-                    data-ng-click="TelecomSmsOptionsManageCtrl.update(TelecomSmsOptionsManageCtrl.service)"
+                    data-ng-if="!TelecomSmsOptionsCallbackCtrl.loading.init"
+                    data-ng-click="TelecomSmsOptionsCallbackCtrl.update(TelecomSmsOptionsCallbackCtrl.service)"
                     data-translate="sms_modify"
                 ></button>
             </div>
@@ -95,4 +95,4 @@
         <!-- /.row -->
     </main>
 </section>
-<!-- /.telecom-sms-options-manage -->
+<!-- /.telecom-sms-options-callback -->
