+++ packages/manager/modules/sms/src/sms/options/callback/translations/Messages_de_DE.json
