+++ packages/manager/modules/sms/src/sms/options/callback/translations/Messages_en_GB.json
