+++ packages/manager/modules/sms/src/sms/options/callback/translations/Messages_fr_FR.json
@@ -1,0 +1,20 @@
+{
+  "sms_section_back_link": "Revenir aux options",
+  "sms_options_callback_title": "Options Callback",
+  "sms_options_callback_information_title": "Informations",
+  "sms_options_callback_information_text": "OVHcloud vous propose de paramétrer des URLs de callback pour être averti lors d’un événement sur un SMS.",
+  "sms_options_callback_information_types_available": "Deux types de callbacks sont disponibles :",
+  "sms_options_callback_information_type_callback": "<b>Le callback de changement d’état</b> : Lorsqu’un SMS envoyé change d’état (ex: de \"envoyé\" vers \"remis\").",
+  "sms_options_callback_information_type_stop_callback": "<b>Le callback STOP</b> : Lorsqu’un destinataire répond \"STOP\" à un de vos SMS pour vous blacklister.",
+  "sms_options_callback_configure_title": "Configurer",
+  "sms_options_callback_configure_callback_label": "URL de callback changements d’état :",
+  "sms_options_callback_configure_stop_callback_label": "URL de callback STOP :",
+  "sms_options_callback_configure_callback_empty": "Aucune URL configurée.",
+  "sms_options_callback_update_title": "Configuration des URLs",
+  "sms_options_callback_update_callback_label": "URL de callback changements d’état",
+  "sms_options_callback_update_stopcallback_label": "URL de callback STOP",
+  "sms_options_callback_update_callbacks_error_pattern": "Le lien doit commencer par http:// ou https://.",
+  "sms_options_callback_update_callbacks_error_url": "L’URL renseignée n’est pas valide.",
+  "sms_options_callback_updating": "Configuration des URLs…",
+  "sms_options_callback_updated": "Configuration des URLs effectuée"
+}
