+++ packages/manager/modules/sms/src/sms/options/callback/translations/Messages_it_IT.json
