+++ packages/manager/modules/sms/src/sms/options/callback/translations/Messages_pl_PL.json
