+++ packages/manager/modules/sms/src/sms/options/callback/translations/Messages_pt_PT.json
@@ -1,12 +1,12 @@
 {
   "sms_section_back_link": "Revenir aux options",
   "sms_options_manage_title": "Options générales",
-  "sms_options_manage_information_title": "Informations",
+  "sms_options_manage_information_title": "Informações",
   "sms_options_manage_information_text": "OVH vous propose de paramétrer des URLs de callback pour être averti lors d’un événement sur un SMS.",
   "sms_options_manage_information_types_available": "Deux types de callbacks sont disponibles :",
   "sms_options_manage_information_type_callback": "<b>Le callback de changement d’état</b> : Lorsqu’un SMS envoyé change d’état (ex: de \"envoyé\" vers \"remis\").",
   "sms_options_manage_information_type_stop_callback": "<b>Le callback STOP</b> : Lorsqu’un destinataire répond \"STOP\" à un de vos SMS pour vous blacklister.",
-  "sms_options_manage_configure_title": "Configurer",
+  "sms_options_manage_configure_title": "Configurar",
   "sms_options_manage_configure_callback_label": "URL de callback changements d’état :",
   "sms_options_manage_configure_stop_callback_label": "URL de callback STOP :",
   "sms_options_manage_configure_callback_empty": "Aucune URL configurée.",
