+++ packages/manager/modules/sms/src/sms/options/callback/update/telecom-sms-options-callback-update.html
@@ -1,0 +1,143 @@
+<form
+    name="updateCallbackOptionsForm"
+    id="updateCallbackOptionsForm"
+    data-ng-submit="OptionsCallbackUpdateCtrl.setUrls()"
+>
+    <div class="modal-header">
+        <button
+            type="button"
+            class="close float-right"
+            aria-label="{{ ::'sms_cancel' | translate }}"
+            data-ng-click="OptionsCallbackUpdateCtrl.cancel()"
+        >
+            <span aria-hidden="true">&times;</span>
+        </button>
+    </div>
+
+    <div
+        class="modal-body"
+        data-ng-if="!OptionsCallbackUpdateCtrl.loading.updateOptions && !OptionsCallbackUpdateCtrl.updated"
+    >
+        <h3
+            class="modal-title"
+            data-translate="sms_options_callback_update_title"
+        ></h3>
+        <!-- CALLBACK -->
+        <div
+            class="form-group"
+            data-ng-class="{
+                 'has-error': updateCallbackOptionsForm.callBack.$touched && updateCallbackOptionsForm.callBack.$invalid
+             }"
+        >
+            <label
+                for="callBack"
+                class="control-label"
+                data-translate="sms_options_callback_update_callback_label"
+            >
+            </label>
+            <input
+                type="url"
+                name="callBack"
+                id="callBack"
+                class="form-control"
+                data-translate-attr="{ placeholder: 'sms_options_callback_update_callback_label' }"
+                data-ng-pattern="OptionsCallbackUpdateCtrl.urlPattern"
+                data-ng-model="OptionsCallbackUpdateCtrl.model.service.callBack"
+            />
+            <span
+                class="help-block"
+                role="alert"
+                data-ng-if="updateCallbackOptionsForm.callBack.$touched"
+                data-ng-messages="updateCallbackOptionsForm.callBack.$error"
+            >
+                <span
+                    data-ng-message="pattern"
+                    data-translate="sms_options_callback_update_callbacks_error_pattern"
+                >
+                </span>
+                <span
+                    data-ng-message="url"
+                    data-translate="sms_options_callback_update_callbacks_error_url"
+                >
+                </span>
+            </span>
+        </div>
+
+        <!-- STOPCALLBACK -->
+        <div
+            class="form-group"
+            data-ng-class="{
+                'has-error': updateCallbackOptionsForm.stopCallBack.$touched && updateCallbackOptionsForm.stopCallBack.$invalid
+            }"
+        >
+            <label
+                for="stopCallBack"
+                class="control-label"
+                data-translate="sms_options_callback_update_stopcallback_label"
+            >
+            </label>
+            <input
+                type="url"
+                name="stopCallBack"
+                id="stopCallBack"
+                class="form-control"
+                data-translate-attr="{ placeholder: 'sms_options_callback_update_stopcallback_label' }"
+                data-ng-pattern="OptionsCallbackUpdateCtrl.urlPattern"
+                data-ng-model="OptionsCallbackUpdateCtrl.model.service.stopCallBack"
+            />
+            <span
+                class="help-block"
+                role="alert"
+                data-ng-if="updateCallbackOptionsForm.stopCallBack.$touched"
+                data-ng-messages="updateCallbackOptionsForm.stopCallBack.$error"
+            >
+                <span
+                    data-ng-message="pattern"
+                    data-translate="sms_options_callback_update_callbacks_error_pattern"
+                >
+                </span>
+                <span
+                    data-ng-message="url"
+                    data-translate="sms_options_callback_update_callbacks_error_url"
+                >
+                </span>
+            </span>
+        </div>
+    </div>
+
+    <div
+        class="modal-body text-center"
+        data-ng-if="OptionsCallbackUpdateCtrl.loading.updateOptions"
+    >
+        <oui-spinner></oui-spinner>
+        <h4 data-translate="sms_options_callback_updating"></h4>
+    </div>
+
+    <div class="modal-body" data-ng-if="OptionsCallbackUpdateCtrl.updated">
+        <tuc-success-drawing-check
+            data-draw-success-check="OptionsCallbackUpdateCtrl.updated"
+        >
+            <h4 data-translate="sms_options_callback_updated"></h4>
+        </tuc-success-drawing-check>
+    </div>
+
+    <div class="modal-footer">
+        <button
+            type="button"
+            class="btn btn-default"
+            data-ng-click="OptionsCallbackUpdateCtrl.cancel()"
+            data-translate="sms_cancel"
+            data-ng-disabled="OptionsCallbackUpdateCtrl.loading.updateOptions ||
+                              OptionsCallbackUpdateCtrl.updated"
+        ></button>
+        <button
+            type="submit"
+            class="btn btn-primary"
+            data-translate="sms_submit"
+            data-ng-disabled="updateCallbackOptionsForm.$invalid ||
+                              !OptionsCallbackUpdateCtrl.hasChanged() ||
+                              OptionsCallbackUpdateCtrl.loading.updateOptions ||
+                              OptionsCallbackUpdateCtrl.updated"
+        ></button>
+    </div>
+</form>
