--- packages/manager/modules/sms/src/sms/options/manage/update/telecom-sms-options-manage-update.html
@@ -1,143 +1,0 @@
-<form
-    name="updateManageOptionsForm"
-    id="updateManageOptionsForm"
-    data-ng-submit="OptionsManageUpdateCtrl.setUrls()"
->
-    <div class="modal-header">
-        <button
-            type="button"
-            class="close float-right"
-            aria-label="{{ ::'sms_cancel' | translate }}"
-            data-ng-click="OptionsManageUpdateCtrl.cancel()"
-        >
-            <span aria-hidden="true">&times;</span>
-        </button>
-    </div>
-
-    <div
-        class="modal-body"
-        data-ng-if="!OptionsManageUpdateCtrl.loading.updateOptions && !OptionsManageUpdateCtrl.updated"
-    >
-        <h3
-            class="modal-title"
-            data-translate="sms_options_manage_update_title"
-        ></h3>
-        <!-- CALLBACK -->
-        <div
-            class="form-group"
-            data-ng-class="{
-                 'has-error': updateManageOptionsForm.callBack.$touched && updateManageOptionsForm.callBack.$invalid
-             }"
-        >
-            <label
-                for="callBack"
-                class="control-label"
-                data-translate="sms_options_manage_update_callback_label"
-            >
-            </label>
-            <input
-                type="url"
-                name="callBack"
-                id="callBack"
-                class="form-control"
-                data-translate-attr="{ placeholder: 'sms_options_manage_update_callback_label' }"
-                data-ng-pattern="OptionsManageUpdateCtrl.urlPattern"
-                data-ng-model="OptionsManageUpdateCtrl.model.service.callBack"
-            />
-            <span
-                class="help-block"
-                role="alert"
-                data-ng-if="updateManageOptionsForm.callBack.$touched"
-                data-ng-messages="updateManageOptionsForm.callBack.$error"
-            >
-                <span
-                    data-ng-message="pattern"
-                    data-translate="sms_options_manage_update_callbacks_error_pattern"
-                >
-                </span>
-                <span
-                    data-ng-message="url"
-                    data-translate="sms_options_manage_update_callbacks_error_url"
-                >
-                </span>
-            </span>
-        </div>
-
-        <!-- STOPCALLBACK -->
-        <div
-            class="form-group"
-            data-ng-class="{
-                'has-error': updateManageOptionsForm.stopCallBack.$touched && updateManageOptionsForm.stopCallBack.$invalid
-            }"
-        >
-            <label
-                for="stopCallBack"
-                class="control-label"
-                data-translate="sms_options_manage_update_stopcallback_label"
-            >
-            </label>
-            <input
-                type="url"
-                name="stopCallBack"
-                id="stopCallBack"
-                class="form-control"
-                data-translate-attr="{ placeholder: 'sms_options_manage_update_stopcallback_label' }"
-                data-ng-pattern="OptionsManageUpdateCtrl.urlPattern"
-                data-ng-model="OptionsManageUpdateCtrl.model.service.stopCallBack"
-            />
-            <span
-                class="help-block"
-                role="alert"
-                data-ng-if="updateManageOptionsForm.stopCallBack.$touched"
-                data-ng-messages="updateManageOptionsForm.stopCallBack.$error"
-            >
-                <span
-                    data-ng-message="pattern"
-                    data-translate="sms_options_manage_update_callbacks_error_pattern"
-                >
-                </span>
-                <span
-                    data-ng-message="url"
-                    data-translate="sms_options_manage_update_callbacks_error_url"
-                >
-                </span>
-            </span>
-        </div>
-    </div>
-
-    <div
-        class="modal-body text-center"
-        data-ng-if="OptionsManageUpdateCtrl.loading.updateOptions"
-    >
-        <oui-spinner></oui-spinner>
-        <h4 data-translate="sms_options_manage_updating"></h4>
-    </div>
-
-    <div class="modal-body" data-ng-if="OptionsManageUpdateCtrl.updated">
-        <tuc-success-drawing-check
-            data-draw-success-check="OptionsManageUpdateCtrl.updated"
-        >
-            <h4 data-translate="sms_options_manage_updated"></h4>
-        </tuc-success-drawing-check>
-    </div>
-
-    <div class="modal-footer">
-        <button
-            type="button"
-            class="btn btn-default"
-            data-ng-click="OptionsManageUpdateCtrl.cancel()"
-            data-translate="sms_cancel"
-            data-ng-disabled="OptionsManageUpdateCtrl.loading.updateOptions ||
-                              OptionsManageUpdateCtrl.updated"
-        ></button>
-        <button
-            type="submit"
-            class="btn btn-primary"
-            data-translate="sms_submit"
-            data-ng-disabled="updateManageOptionsForm.$invalid ||
-                              !OptionsManageUpdateCtrl.hasChanged() ||
-                              OptionsManageUpdateCtrl.loading.updateOptions ||
-                              OptionsManageUpdateCtrl.updated"
-        ></button>
-    </div>
-</form>
