+++ packages/manager/modules/sms/src/sms/options/translations/Messages_fr_FR.json
@@ -1,6 +1,6 @@
 {
   "sms_options_breadcrumb": "Options",
-  "sms_options_manage": "Options générales",
+  "sms_options_callback": "Options Callback",
   "sms_options_response": "Options des réponses",
   "sms_options_recredit": "Recharge automatique",
   "sms_options_blacklist": "Gérer les expéditeurs blacklistés",
